Intake and Perception/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
Pillow
transformers
torch
numpy
mistralai
requests
python-dotenv